requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.0

# AI
anthropic==0.42.0
voyageai==0.3.2

# OpenAPI parsing
pyyaml==6.0.2
jsonschema==4.23.0

# HTTP client (for API gateway sync)
httpx==0.28.1

# Utilities
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
